+++ libs/server-sent-events/src/client.cpp
@@ -79,31 +79,38 @@ class FoxyClient : public Client,
           read_timeout_(read_timeout),
           write_timeout_(write_timeout),
           req_(std::move(req)),
-          session_(std::move(executor),
-                   launchdarkly::foxy::session_opts{
-                       ToOptRef(ssl_context_),
-                       connect_timeout.value_or(kNoTimeout)}),
+          event_receiver_(std::move(receiver)),
+          logger_(std::move(logger)),
+          errors_(std::move(errors)),
+          body_parser_(std::nullopt),
+          session_(std::nullopt),
+          last_event_id_(std::nullopt),
           backoff_(
               initial_reconnect_delay.value_or(kDefaultInitialReconnectDelay),
               kDefaultMaxBackoffDelay),
-          last_event_id_(std::nullopt),
-          backoff_timer_(session_.get_executor()),
-          event_receiver_(std::move(receiver)),
-          logger_(std::move(logger)),
-          errors_(std::move(errors)) {
+          backoff_timer_(std::move(executor)),
+          last_read_(std::nullopt),
+          shutting_down_(false) {
+        create_session();
         create_parser();
     }
 
-    /** The body parser is recreated each time a connection is made because
-     * its internal state cannot be explicitly reset.
-     *
-     * Since SSE body will never end unless
-     * an error occurs, the body size limit must be removed.
-     */
+    // The body parser is recreated each time a connection is made because
+    // its internal state cannot be explicitly reset.
     void create_parser() {
         body_parser_.emplace();
+        // Remove body read limit because an SSE stream can be infinite.
         body_parser_->body_limit(boost::none);
         body_parser_->get().body().on_event(event_receiver_);
+    }
+
+    // The session is recreated each time a connection is made because its
+    // internal state cannot be explicitly reset.
+    void create_session() {
+        session_.emplace(
+            backoff_timer_.get_executor(),
+            launchdarkly::foxy::session_opts{
+                ToOptRef(ssl_context_), connect_timeout_.value_or(kNoTimeout)});
     }
 
     /**
@@ -113,7 +120,7 @@ class FoxyClient : public Client,
      * The body parser's last SSE event ID must be cached so it can be added
      * as a header on the next request (since the parser is destroyed.)
      */
-    void do_backoff(std::string const& reason) {
+    void async_backoff(std::string const& reason) {
         backoff_.fail();
 
         if (auto id = body_parser_->get().body().last_event_id()) {
@@ -131,6 +138,7 @@ class FoxyClient : public Client,
 
         logger_(msg.str());
 
+        create_session();
         create_parser();
         backoff_timer_.expires_from_now(backoff_.delay());
         backoff_timer_.async_wait(beast::bind_front_handler(
@@ -141,11 +149,17 @@ class FoxyClient : public Client,
         if (ec == boost::asio::error::operation_aborted) {
             return;
         }
-        run();
+        do_run();
     }
 
-    void run() override {
-        session_.async_connect(
+    void async_connect() override {
+        boost::asio::post(
+            session_->get_executor(),
+            beast::bind_front_handler(&FoxyClient::do_run, shared_from_this()));
+    }
+
+    void do_run() {
+        session_->async_connect(
             host_, port_,
             beast::bind_front_handler(&FoxyClient::on_connect,
                                       shared_from_this()));
@@ -156,7 +170,7 @@ class FoxyClient : public Client,
             return;
         }
         if (ec) {
-            return do_backoff(ec.what());
+            return async_backoff(ec.what());
         }
 
         if (last_event_id_) {
@@ -164,10 +178,10 @@ class FoxyClient : public Client,
         } else {
             req_.erase("last-event-id");
         }
-        session_.opts.timeout = write_timeout_.value_or(kNoTimeout);
-        session_.async_write(req_,
-                             beast::bind_front_handler(&FoxyClient::on_write,
-                                                       shared_from_this()));
+        session_->opts.timeout = write_timeout_.value_or(kNoTimeout);
+        session_->async_write(req_,
+                              beast::bind_front_handler(&FoxyClient::on_write,
+                                                        shared_from_this()));
     }
 
     void on_write(boost::system::error_code ec, std::size_t amount) {
@@ -176,11 +190,11 @@ class FoxyClient : public Client,
             return;
         }
         if (ec) {
-            return do_backoff(ec.what());
+            return async_backoff(ec.what());
         }
 
-        session_.opts.timeout = read_timeout_.value_or(kNoTimeout);
-        session_.async_read_header(
+        session_->opts.timeout = read_timeout_.value_or(kNoTimeout);
+        session_->async_read_header(
             *body_parser_, beast::bind_front_handler(&FoxyClient::on_headers,
                                                      shared_from_this()));
     }
@@ -191,12 +205,12 @@ class FoxyClient : public Client,
             return;
         }
         if (ec) {
-            return do_backoff(ec.what());
+            return async_backoff(ec.what());
         }
 
         if (!body_parser_->is_header_done()) {
             /* keep reading headers */
-            return session_.async_read_header(
+            return session_->async_read_header(
                 *body_parser_,
                 beast::bind_front_handler(&FoxyClient::on_headers,
                                           shared_from_this()));
@@ -212,11 +226,14 @@ class FoxyClient : public Client,
                 return;
             }
             if (!correct_content_type(response)) {
-                return do_backoff("invalid Content-Type");
+                return async_backoff("invalid Content-Type");
             }
 
+            logger_("connected");
             backoff_.succeed();
-            return session_.async_read(
+
+            last_read_ = std::chrono::steady_clock::now();
+            return session_->async_read_some(
                 *body_parser_,
                 beast::bind_front_handler(&FoxyClient::on_read_body,
                                           shared_from_this()));
@@ -244,14 +261,14 @@ class FoxyClient : public Client,
 
         if (status_class == beast::http::status_class::client_error) {
             if (recoverable_client_error(response.result())) {
-                return do_backoff(backoff_reason(response.result()));
+                return async_backoff(backoff_reason(response.result()));
             }
 
             errors_(Error::UnrecoverableClientError);
             return;
         }
 
-        do_backoff(backoff_reason(response.result()));
+        async_backoff(backoff_reason(response.result()));
     }
 
     static std::string backoff_reason(beast::http::status status) {
@@ -262,38 +279,93 @@ class FoxyClient : public Client,
 
     void on_read_body(boost::system::error_code ec, std::size_t amount) {
         boost::ignore_unused(amount);
-        if (ec == boost::asio::error::operation_aborted) {
-            return;
+        if (ec) {
+            if (ec == boost::asio::error::operation_aborted) {
+                // operation_aborted can occur if the read timeout is reached or
+                // if we're shutting down, so shutting_down_ is needed to
+                // disambiguate.
+                if (shutting_down_) {
+                    // End the chain of async operations.
+                    return;
+                }
+                errors_(Error::ReadTimeout);
+                return async_backoff(
+                    "aborting read of response body (timeout)");
+            }
+            return async_backoff(ec.what());
         }
-        do_backoff(ec.what());
+
+        // The server can indicate that the chunk encoded response is done
+        // by sending a final chunk + CRLF. The body parser will have
+        // detected this.
+        if (body_parser_->is_done()) {
+            return async_backoff("receiving final chunk");
+        }
+
+        log_and_update_last_read(amount);
+        return session_->async_read_some(
+            *body_parser_, beast::bind_front_handler(&FoxyClient::on_read_body,
+                                                     shared_from_this()));
+    }
+
+    /** Logs a message indicating that an async_read_some operation
+     * on the response's body has completed, and how long that operation took
+     * (using a monotonic clock.)
+     *
+     * This is useful for debugging timeout-related issues, like receiving a
+     * heartbeat message.
+     */
+    void log_and_update_last_read(std::size_t amount) {
+        if (last_read_) {
+            auto sec_since_last_read =
+                std::chrono::duration_cast<std::chrono::seconds>(
+                    std::chrono::steady_clock::now() - *last_read_)
+                    .count();
+            logger_("read (" + std::to_string(amount) + ") bytes in (" +
+                    std::to_string(sec_since_last_read) + ") sec");
+        } else {
+            logger_("read (" + std::to_string(amount) + ") bytes");
+        }
+        last_read_ = std::chrono::steady_clock::now();
     }
 
     void async_shutdown(std::function<void()> completion) override {
-        boost::asio::post(session_.get_executor(),
+        // Get on the session's executor, otherwise the code in the completion
+        // handler could race.
+        boost::asio::post(session_->get_executor(),
                           beast::bind_front_handler(&FoxyClient::do_shutdown,
                                                     shared_from_this(),
                                                     std::move(completion)));
     }
 
     void do_shutdown(std::function<void()> completion) {
-        session_.async_shutdown(beast::bind_front_handler(
+        shutting_down_ = true;
+        backoff_timer_.cancel();
+        session_->async_shutdown(beast::bind_front_handler(
             &FoxyClient::on_shutdown, std::move(completion)));
     }
 
     static void on_shutdown(std::function<void()> completion,
                             boost::system::error_code ec) {
+        // Because do_shutdown doesn't use shared_from_this() when initiating
+        // the async_shutdown op, the client may already be destroyed - hence
+        // this static method.
         boost::ignore_unused(ec);
         if (completion) {
             completion();
         }
     }
 
+    // Some client errors are considered recoverable, meaning they could be
+    // resolved by reconnecting. Returns true if the status is one of those.
     static bool recoverable_client_error(beast::http::status status) {
         return (status == beast::http::status::bad_request ||
                 status == beast::http::status::request_timeout ||
                 status == beast::http::status::too_many_requests);
     }
 
+    // A naive comparison of content-type to 'text/event-stream' is incorrect
+    // because multiple content types may be present.
     static bool correct_content_type(FoxyClient::response const& response) {
         if (auto content_type = response.find("content-type");
             content_type != response.end()) {
@@ -303,12 +375,16 @@ class FoxyClient : public Client,
         return false;
     }
 
+    // If the server redirects, ensure the location header is present.
     static bool can_redirect(FoxyClient::response const& response) {
         return (response.result() == beast::http::status::moved_permanently ||
                 response.result() == beast::http::status::temporary_redirect) &&
                response.find("location") != response.end();
     }
 
+    // Generates a redirect URL from a base and provided location. Since the
+    // location might not be absolute, it may be necessary to resolve it against
+    // the base.
     static std::optional<boost::urls::url> redirect_url(
         std::string orig_base,
         std::string orig_location) {
@@ -330,25 +406,77 @@ class FoxyClient : public Client,
     }
 
    private:
+    // Optional, but necessary for establishing TLS connections. If it is
+    // omitted and the scheme is https://, the connection will fail. Can be
+    // reused across connections.
     std::optional<net::ssl::context> ssl_context_;
+
     std::string host_;
     std::string port_;
 
+    // If present, the max amount of time that can be spent resolving DNS
+    // and setting up the initial connection. This doesn't include writing the
+    // request nor receiving the response.
     std::optional<std::chrono::milliseconds> connect_timeout_;
+
+    // If present, the max amount of time that can be spent after reading bytes
+    // before more bytes must be read. Applies after sending the request, but
+    // before receiving the response headers. Intended to terminate connections
+    // where the sender has hung up. LaunchDarkly sends a heartbeat every 180
+    // seconds, so the read timeout must be greater than this.
     std::optional<std::chrono::milliseconds> read_timeout_;
+
+    // If present, the max amount of time that can be spent sending the request.
     std::optional<std::chrono::milliseconds> write_timeout_;
 
+    // The request sent to the server, which persists across reconnection
+    // attempts.
     http::request<http::string_body> req_;
 
+    // Callback which executed whenever an SSE event is received. This is
+    // passed into the body_parser whenever it is constructed, so it needs to be
+    // stored here for future use.
     Builder::EventReceiver event_receiver_;
+
+    // Callback executed when log messages must be generated. The provider of
+    // the callback dictates the log level, which at the moment of writing is
+    // 'debug'.
     Builder::LogCallback logger_;
+
+    // Callback executed to report errors. This is the primary mechanism by
+    // which the client communicates error conditions to the user.
     Builder::ErrorCallback errors_;
 
+    // Customized parser (see parser.hpp) which repeatedly receives chunks of
+    // data and parses them into SSE events. It cannot be reused across
+    // connections, hence the optional so it can be destroyed easily.
     std::optional<http::response_parser<body>> body_parser_;
-    launchdarkly::foxy::client_session session_;
+
+    // The primary network primitive used to read/write to the server. It is our
+    // responsibility to ensure it is called from only one thread at a time. It
+    // cannot be reused across connections, similar to the body_parser.
+    std::optional<launchdarkly::foxy::client_session> session_;
+
+    // Stores the last known SSE event ID, which can be provided to the server
+    // upon reconnection.
     std::optional<std::string> last_event_id_;
+
+    // Computes the backoff delays used whenever the connection drops
+    // unnaturally.
     Backoff backoff_;
+
+    // Used in concert with the backoff member to delay reconnection attempts.
+    // The timer must be cancelled when shutdown is requested, because an
+    // outstanding backoff delay might be in progress.
     boost::asio::steady_timer backoff_timer_;
+
+    // Stores the timestamp of the last read in order to augment debug logging.
+    std::optional<std::chrono::steady_clock::time_point> last_read_;
+
+    // Upon completing a read, a new async read is generally initiated. This
+    // determines whether to return and thus end the operation gracefully (true)
+    // or to perform backoff (false).
+    bool shutting_down_;
 };
 
 Builder::Builder(net::any_io_executor ctx, std::string url)